INS-245: moved message struct from message router to service. removed cyclomatic import
